requirements.txt
PySide6>=6.4.0
PyInstaller>=5.6.0
Pillow>=9.0.0
matplotlib>=3.5.0
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.1
requests==2.31.0
PyPDF2==3.0.0
pandas==2.0.3
termcolor==2.4.0
numpy==1.24.3
pandas-ta==0.3.14b
schedule==1.2.0
python-dateutil==2.8.2
backoff==2.2.1
youtube-transcript-api==0.6.2
openai==1.59.5
groq==0.16.0  # Moon Dev's Groq API integration 🚀
selenium>=4.16.0
pillow>=10.2.0
webdriver-manager>=4.0.1
# Add any other dependencies your agents need
google-api-python-client==2.107.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
# YouTube Clipper Agent dependencies
pytube==15.0.0
ffmpeg-python==0.2.0
solders==0.26.0
solana==0.36.6
anthropic==0.18.1
aiohttp==3.8.4
